Properly await running index write during dispose
DiskStore#dispose only attempted to wait for the currently scheduled index write if it's running,
or cancelling it otherwise so it won't start later. This wasn't enough, since the currently running
task might be lagging behind the currently scheduled, maybe by multiple tasks. Successfully cancelling
the currently scheduled write doesn't guarantee no writes are running or are about to run later. This
caused scarce NoSuchFileExceptions in tests due to a concurrent index update while the store was disposing.
Using a Phaser with 1 static registerer (the scheduler) & 1 dynamic registerer (the currently running write)
solves this issue.

The change also introduces a Delayer indirection in place of directly using CompletableFuture#delayedExecutor.
This, together with the use of a simple MockClock listener, allows tests to compute delays in a controlled and
hence deterministic environment.

Also, StoreProvider now supports both ArgumentsProvider & ParameterResolver interfaces.