Upgrade gradle (to 7.0) & dependencies
Relevant deps for changelog:
- Jackson: 2.11.2 -> 2.12.3
- Reactor: 3.4.0 -> 3.4.5
- Errorprone (compile-time only): 2.4.0 -> 2.6.0

OkHttp wasn't upgraded as the newest version bombarded the build with
'warning: unknown enum constant DeprecationLevel.ERROR' for some unknown reason. It also
caused the build to fail with 'class file for kotlin.jvm.internal.markers.KMappedMarker not found'
for some other unknown reason. It's now tempting to fork MockWebServer, rewrite it in Java and dump
okhttp test dependency to not have to deal with this again.