Fix missed response for HttpResponsePublisher (#69)
Because `onResponse(response)` decrements `ongoing` prematurely (before
submitting the response and potentially completing downstream), the last
response passed might be missed when a race happens with
`onInitialResponseBodyCompletion()`. Consider this scenario:

 - `onResponse` & `onInitialResponseBodyCompletion` interleave.
 - `onResponse` acquires the lock first, and decrements `ongoing`. It
   won't perceive initial body completion, so it won't want to complete
   downstream.
 - `onInitialResponseBodyCompletion` acquires the lock later, and perceives
   no ongoing responses, so it proceeds with the intent to complete downstream.
 - `onInitialResponseBodyCompletion` continues, and calls `complete()`, which
   pushes the sentinel to the queue.
 - `onResponse` continues to `submit(response)`, but the response is submitted
   after the sentinel and is hence ignored.

The fix is to not prematurely decrement `ongoing` when the last response is received,
but do so after the response is submitted. This way, `onInitialResponseBodyCompletion`
never perceives a completed stream that is not **really** completed.