Optionally wait for committing the edit before completing downstream
For later reference, here's the proof of why this works. We're interested in proving that
the publisher eventually completes downstream exactly once. We'll take into consideration three
methods in CacheWritingSubscription and one fact.
 - CacheWritingSubscription::onComplete
 - CacheWritingSubscription::completeDownstreamOnCommittedEdit
 - CacheWritingSubscription::completeDownstreamOnDiscardedEdit
 - CacheWritingSubscription::onComplete AND exactly one of CacheWritingSubscription::completeDownstreamOnCommittedEdit
   or CacheWritingSubscription::completeDownstreamOnDiscardedEdit are eventually called.

When waitForCommit is true, downstream is completed immediately when completion comes from upstream. When
waitForCommit is false and the edit is committed, downstream is completed (by completeDownstreamONCommittedEdit)
as soon as the edit is committed. Proof of correctness trivially holds here. Now let's see what happens when
waitForCommit is false and the edit is discarded due to error. This is a trickier case as there's no deterministic
order to calling CacheWritingSubscription::discardEdit & CacheWritingSubscription::onComplete in this case
(e.g. if a write falls, discardEdit can be called anytime, before, after or concurrently with onComplete). This means
that whichever of them runs "later" is responsible for completing downstream. Supposing T1 calls onComplete and
T2 calls discardEdit then completeDownstreamOnDiscardedEdit, the implementation ensures that:
 - write[T1](receivedBodyCompletion = true) -> read[T1](state == DONE) -> completeDownstream[T1]()
or
 - write[T2](state == DONE) -> read[T2](responseBodyCompletion == true)  -> completeDownstream[T2]()

Where "write[T2](state == DONE) -> read[T2](responseBodyCompletion == true)" means T2 reading state as DONE happens-before
T2 writing receivedBodyCompletion to true. Let's assume that this doesn't work. That is, neither of them call
completeDownstream(). This means that neither T1 does read(state == DONE) nor T2 does
read(responseBodyCompletion == true). In other words:
 - write[T1](receivedBodyCompletion = true) -> read[T1](state != DONE)
and
 - write[T2](state == DONE) -> read[T2](responseBodyCompletion != true)

But notice that each thread sets the value that the other method didn't see. Thus, each thread not reading that value must
happen-before the other thread writing it. Now by joining:
 - write[T1](receivedBodyCompletion = true) -> read[T1](state != DONE) -> write[T2](state == DONE) -> read[T2](responseBodyCompletion != true)
and
 - write[T2](state == DONE) -> read[T2](responseBodyCompletion != true) -> write[T1](receivedBodyCompletion = true) -> read[T1](state != DONE)

This means that:
 - write[T1](receivedBodyCompletion = true) -> read[T2](responseBodyCompletion != true)
and
 - write[T2](state == DONE) -> read[T1](state != DONE)

Which is a contradiction, since responseBodyCompletion is never written to false after being true, and state is never written to
anything after being DONE. Thus, at least T1 or T2 must've completed downstream. The fact that it is completed once, however,
relies on the fact that the completing thread atomically unsets downstream to null before completing, so it becomes to other
threads.